+++ lib/rblade/compiler/compiles_ruby.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module RBlade
   class CompilesRuby
     def compile!(tokens)