+++ lib/rblade/compiler/tokenizes_statements.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "rblade/helpers/tokenizer"
 require "ripper"