+++ lib/rblade/helpers/class_manager.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module RBlade
   class ClassManager
     def initialize classes
@@ -6,7 +8,7 @@ module RBlade
       elsif classes.is_a? Array
         @classes = classes.join " "
       elsif classes.is_a? Hash
-        @classes = ""
+        @classes = +""
         classes.map do |value, predicate|
           if predicate
             @classes << "#{value} "