+++ test/unit/compiler/concerns/tokenizes_statements_test.rb
@@ -21,58 +21,58 @@ class TokenizesStatementsTest < TestCase
   end
 
   def test_tokenize
-    assert_tokenizes_to "@foo", [{name: "foo"}]
-    assert_tokenizes_to "@foo @bar", [{name: "foo"}, {name: "bar"}]
-    assert_tokenizes_to "@foo  @bar", [{name: "foo"}, {name: "bar"}]
-    assert_tokenizes_to "@foo   @bar", [{name: "foo"}, " ", {name: "bar"}]
-    assert_tokenizes_to "@foo()", [{name: "foo"}]
-    assert_tokenizes_to "@foo(1, 2, 3)", [{name: "foo", arguments: ["1", "2", "3"]}]
-    assert_tokenizes_to "@foo   (1, 2, 3)", [{name: "foo", arguments: ["1", "2", "3"]}]
-    assert_tokenizes_to "@foo   (1,
+    assert_tokenizes_to "@ruby", [{name: "ruby"}]
+    assert_tokenizes_to "@ruby @if", [{name: "ruby"}, {name: "if"}]
+    assert_tokenizes_to "@ruby  @if", [{name: "ruby"}, {name: "if"}]
+    assert_tokenizes_to "@ruby   @if", [{name: "ruby"}, " ", {name: "if"}]
+    assert_tokenizes_to "@ruby()", [{name: "ruby"}]
+    assert_tokenizes_to "@ruby(1, 2, 3)", [{name: "ruby", arguments: ["1", "2", "3"]}]
+    assert_tokenizes_to "@ruby   (1, 2, 3)", [{name: "ruby", arguments: ["1", "2", "3"]}]
+    assert_tokenizes_to "@ruby   (1,
     2,
-    3)", [{name: "foo", arguments: ["1", "2", "3"]}]
+    3)", [{name: "ruby", arguments: ["1", "2", "3"]}]
   end
 
   def test_nested_statements
-    assert_tokenizes_to "@foo(@bar)", [{name: "foo", arguments: ["@bar"]}]
-    assert_tokenizes_to "@foo(@@bar)", [{name: "foo", arguments: ["@@bar"]}]
-    assert_tokenizes_to "@@foo(@bar)", ["@foo", "(", {name: "bar"}, ")"]
-    assert_tokenizes_to "@@foo(()@bar)", ["@foo", "(()", {name: "bar"}, ")"]
+    assert_tokenizes_to "@ruby(@if)", [{name: "ruby", arguments: ["@if"]}]
+    assert_tokenizes_to "@ruby(@@if)", [{name: "ruby", arguments: ["@@if"]}]
+    assert_tokenizes_to "@@ruby(@if)", ["@ruby", "(", {name: "if"}, ")"]
+    assert_tokenizes_to "@@ruby(()@if)", ["@ruby", "(()", {name: "if"}, ")"]
   end
 
   def test_skip_statement
-    assert_tokenizes_to "@@foo", ["@foo"]
-    assert_tokenizes_to "@@foo(1, 2, 3)", ["@foo", "(1, 2, 3)"]
+    assert_tokenizes_to "@@ruby", ["@ruby"]
+    assert_tokenizes_to "@@ruby(1, 2, 3)", ["@ruby", "(1, 2, 3)"]
   end
 
   def test_bracket_matching
-    assert_tokenizes_to "@foo()", [{name: "foo"}]
-    assert_tokenizes_to "@foo(')', 2)", [{name: "foo", arguments: ["')'", "2"]}]
-    assert_tokenizes_to "@foo('(', 2)", [{name: "foo", arguments: ["'('", "2"]}]
-    assert_tokenizes_to "@foo(%q[)], 2)", [{name: "foo", arguments: ["%q[)]", "2"]}]
+    assert_tokenizes_to "@ruby()", [{name: "ruby"}]
+    assert_tokenizes_to "@ruby(')', 2)", [{name: "ruby", arguments: ["')'", "2"]}]
+    assert_tokenizes_to "@ruby('(', 2)", [{name: "ruby", arguments: ["'('", "2"]}]
+    assert_tokenizes_to "@ruby(%q[)], 2)", [{name: "ruby", arguments: ["%q[)]", "2"]}]
 
-    assert_tokenizes_to "@foo((1), ((2)), (3))", [{name: "foo", arguments: ["(1)", "((2))", "(3)"]}]
-    assert_tokenizes_to "@foo((1 + 2), 3)", [{name: "foo", arguments: ["(1 + 2)", "3"]}]
-    assert_tokenizes_to "@foo(#{"(" * 100}#{")" * 100})", [{name: "foo", arguments: ["#{"(" * 100}#{")" * 100}"]}]
+    assert_tokenizes_to "@ruby((1), ((2)), (3))", [{name: "ruby", arguments: ["(1)", "((2))", "(3)"]}]
+    assert_tokenizes_to "@ruby((1 + 2), 3)", [{name: "ruby", arguments: ["(1 + 2)", "3"]}]
+    assert_tokenizes_to "@ruby(#{"(" * 100}#{")" * 100})", [{name: "ruby", arguments: ["#{"(" * 100}#{")" * 100}"]}]
 
-    assert_tokenizes_to "@foo(1,
-    2)", [{name: "foo", arguments: ["1", "2"]}]
-    assert_tokenizes_to "@foo(1
-, 2)", [{name: "foo", arguments: ["1", "2"]}]
-    assert_tokenizes_to "@foo(1,
+    assert_tokenizes_to "@ruby(1,
+    2)", [{name: "ruby", arguments: ["1", "2"]}]
+    assert_tokenizes_to "@ruby(1
+, 2)", [{name: "ruby", arguments: ["1", "2"]}]
+    assert_tokenizes_to "@ruby(1,
     (2 + (3)),
-    4) @bar()", [{name: "foo", arguments: ["1", "(2 + (3))", "4"]}, {name: "bar"}]
+    4) @if()", [{name: "ruby", arguments: ["1", "(2 + (3))", "4"]}, {name: "if"}]
 
-    assert_tokenizes_to "( @foo())", ["(", {name: "foo"}, ")"]
-    assert_tokenizes_to "@foo)", [{name: "foo"}, ")"]
-    assert_tokenizes_to "@foo(", [{name: "foo"}, "("]
-    assert_tokenizes_to "@foo(()(()", [{name: "foo"}, "(()(()"]
+    assert_tokenizes_to "( @ruby())", ["(", {name: "ruby"}, ")"]
+    assert_tokenizes_to "@ruby)", [{name: "ruby"}, ")"]
+    assert_tokenizes_to "@ruby(", [{name: "ruby"}, "("]
+    assert_tokenizes_to "@ruby(()(()", [{name: "ruby"}, "(()(()"]
   end
 
   def test_commas_in_brackets
-    assert_tokenizes_to "@foo([1, 2, 3])", [{name: "foo", arguments: ["[1, 2, 3]"]}]
-    assert_tokenizes_to "@foo([1, {2, 3}])", [{name: "foo", arguments: ["[1, {2, 3}]"]}]
-    assert_tokenizes_to "@foo([1], ([{2, 3}]))", [{name: "foo", arguments: ["[1]", "([{2, 3}])"]}]
+    assert_tokenizes_to "@ruby([1, 2, 3])", [{name: "ruby", arguments: ["[1, 2, 3]"]}]
+    assert_tokenizes_to "@ruby([1, {2, 3}])", [{name: "ruby", arguments: ["[1, {2, 3}]"]}]
+    assert_tokenizes_to "@ruby([1], ([{2, 3}]))", [{name: "ruby", arguments: ["[1]", "([{2, 3}])"]}]
   end
 
   def test_boundaries