+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.0 [UNRELEASED]
 - Add support for helper functions that use block content (#14)
+- Change to use frozen string literals
 - Fix issue with no space after starting ERB style tag
 
 ## 1.1.1 [2024-08-23]